+++ lib/runners/schema/options.rb
@@ -6,17 +6,16 @@ module Runners
       let :source, object?(
         head: string,
         base: string?,
-        git_http_url: string,
-        git_http_userinfo: string?,
-        owner: string,
-        repo: string,
-        pull_number: number?,
+        git_url: string,
+        git_url_userinfo: string?,
+        refspec: enum?(string, array(string)),
       )
 
       let :payload, object(
         source: source,
         outputs: array?(string),
         ssh_key: string?,
+        s3: object?(endpoint: string),
       )
     end
   end