+++ CHANGELOG.md
@@ -4,7 +4,25 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.33.0...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.34.1...HEAD)
+
+- **Cppcheck** Run multiple threads with the `-j` option [#1462](https://github.com/sider/runners/pull/1462)
+
+## 0.34.1
+
+[Full diff](https://github.com/sider/runners/compare/0.34.0...0.34.1)
+
+- Reconsider `Results::Failure` message [#1459](https://github.com/sider/runners/pull/1459)
+
+## 0.34.0
+
+[Full diff](https://github.com/sider/runners/compare/0.33.0...0.34.0)
+
+Breaking changes:
+
+- Remove :owner and :repo, and rename :git_http_url and :git_http_userinfo [#1411](https://github.com/sider/runners/pull/1411)
+- Replace :pull_number with :refspec [#1412](https://github.com/sider/runners/pull/1412)
+- Unify `S3_ENDPOINT` to `RUNNERS_OPTIONS` [#1414](https://github.com/sider/runners/pull/1414)
 
 ## 0.33.0